We're not gonna need rake
Both rubocop and rspec have cli... so we're not gonna reinvent the
wheel. And of course we're not gonna use rake anymore as long as it's
not necessary.

This is probably a limitation but I'm quite sure it will be a good
exercise from which we will learn and enjoy :)